+++ database/migrations/2022_09_23_195728_create_authorized_posts_table.php
@@ -16,9 +16,9 @@ class CreateAuthorizedPostsTable extends Migration
         Schema::create('authorized_posts', function (Blueprint $table) {
             //$table->id();
             $table->primary(['company_id','branch_id','jop_position_id']);
-            $table->unsignedSmallInteger('company_id')->default(0);
-            $table->unsignedSmallInteger('branch_id')->default(0);
-            $table->unsignedSmallInteger('jop_position_id')->default(0);
+            $table->unsignedInteger('company_id')->default(0);
+            $table->unsignedInteger('branch_id')->default(0);
+            $table->unsignedInteger('jop_position_id')->default(0);
             //$table->unsignedSmallInteger('department_id')->default(0);            
             $table->integer('quantity');
             $table->timestamps();
