+++ database/migrations/2022_10_14_141446_create_candidates_table.php
@@ -18,10 +18,10 @@ class CreateCandidatesTable extends Migration
             $table->string('name');
             $table->string('email')->nullable();                                    
             $table->string('mobile_phone')->nullable();   
-            $table->unsignedSmallInteger('requisition_id')->default(0);
-            $table->unsignedSmallInteger('sources_id')->default(0);         
-            $table->unsignedSmallInteger('status_id')->default(1);
-            $table->unsignedSmallInteger('user_id')->default(0);            
+            $table->unsignedInteger('requisition_id')->default(0);
+            $table->unsignedInteger('sources_id')->default(0);         
+            $table->unsignedInteger('status_id')->default(1);
+            $table->unsignedInteger('user_id')->default(0);            
             $table->string('commentaries',500)->nullable();
             $table->tinyInteger('is_hired')->default(0);
             $table->tinyInteger('is_accepted')->default(0);
